VivaLaVida

+++ fvckximi.sh
@@ -4,27 +4,27 @@ repS="python3 $dir/bin/strRep.py"
 
 jar_util() 
 {
-	cd $dir
-	#binary
-	if [[ $3 == "fw" ]]; then 
-		bak="java -jar $dir/bin/baksmali.jar d"
-		sma="java -jar $dir/bin/smali.jar a"
-	else
-		bak="java -jar $dir/bin/baksmali-2.3.4.jar d"
-		sma="java -jar $dir/bin/smali-2.3.4.jar a"
+
+	if [[ ! -d $dir/jar_temp ]]; then
+		mkdir $dir/jar_temp
 	fi
+
+	#binary
+	bak="java -jar $dir/bin/baksmali.jar d"
+	sma="java -jar $dir/bin/smali.jar a"
 
 	if [[ $1 == "d" ]]; then
 		echo -ne "====> Patching $2 : "
-		if [[ -f $dir/framework.jar ]]; then
-			sudo cp $dir/framework.jar $dir/jar_temp
+
+		if [[ $(get_file_dir $2 ) ]]; then
+			sudo mv $(get_file_dir $2 ) $dir/jar_temp
 			sudo chown $(whoami) $dir/jar_temp/$2
 			unzip $dir/jar_temp/$2 -d $dir/jar_temp/$2.out  >/dev/null 2>&1
 			if [[ -d $dir/jar_temp/"$2.out" ]]; then
 				rm -rf $dir/jar_temp/$2
-				for dex in $(find $dir/jar_temp/"$2.out" -maxdepth 1 -name "*dex" ); do
+				for dex in $(sudo find $dir/jar_temp/"$2.out" -maxdepth 1 -name "*dex" ); do
 						if [[ $4 ]]; then
-							if [[ ! "$dex" == *"$4"* ]]; then
+							if [[ "$dex" != *"$4"* && "$dex" != *"$5"* ]]; then
 								$bak $dex -o "$dex.out"
 								[[ -d "$dex.out" ]] && rm -rf $dex
 							fi
@@ -40,9 +40,10 @@ jar_util()
 		if [[ $1 == "a" ]]; then 
 			if [[ -d $dir/jar_temp/$2.out ]]; then
 				cd $dir/jar_temp/$2.out
-				for fld in $(find -maxdepth 1 -name "*.out" ); do
+				for fld in $(sudo find -maxdepth 1 -name "*.out" ); do
 					if [[ $4 ]]; then
-						if [[ ! "$fld" == *"$4"* ]]; then
+						if [[ "$fld" != *"$4"* && "$fld" != *"$5"* ]]; then
+							echo $fld
 							$sma $fld -o $(echo ${fld//.out})
 							[[ -f $(echo ${fld//.out}) ]] && rm -rf $fld
 						fi
@@ -53,13 +54,11 @@ jar_util()
 				done
 				7za a -tzip -mx=0 $dir/jar_temp/$2_notal $dir/jar_temp/$2.out/. >/dev/null 2>&1
 				#zip -r -j -0 $dir/jar_temp/$2_notal $dir/jar_temp/$2.out/.
-				zipalign 4 $dir/jar_temp/$2_notal $dir/jar_temp/$2
+				zipalign -p -v 4 $dir/jar_temp/$2_notal $dir/jar_temp/$2 >/dev/null 2>&1
 				if [[ -f $dir/jar_temp/$2 ]]; then
-					sudo cp -rf $dir/jar_temp/$2 $dir/module/system/framework
-					final_dir="$dir/module/*"
-					#7za a -tzip "$dir/services_patched_$(date "+%d%m%y").zip" $final_dir
-					echo "Success"
 					rm -rf $dir/jar_temp/$2.out $dir/jar_temp/$2_notal 
+					#sudo cp -rf $dir/jar_temp/$2 $(get_file_dir $2) 
+					echo "Succes"
 				else
 					echo "Fail"
 				fi
@@ -73,7 +72,7 @@ framework() {
 
 	lang_dir="$dir/module/lang"
 
-	jar_util d "framework.jar" fw 
+	jar_util d "framework.jar" fw classes2
 
 	#patch
